🚚 move all 'update' types to the same file, events.ts
