+++ packages/atom.io/src/main/index.ts
@@ -1,5 +1,6 @@
 export * from "./atom"
 export * from "./dispose-state"
+export * from "./events"
 export * from "./find-state"
 export * from "./get-state"
 export * from "./join"
