+++ packages/atom.io/src/main/timeline.ts
@@ -1,16 +1,12 @@
-import type {
-	Timeline,
-	TimelineAtomUpdate,
-	TimelineMoleculeCreation,
-	TimelineMoleculeDisposal,
-	TimelineSelectorUpdate,
-	TimelineStateCreation,
-	TimelineStateDisposal,
-	TimelineTransactionUpdate,
-} from "atom.io/internal"
+import type { Timeline } from "atom.io/internal"
 import { createTimeline, IMPLICIT, timeTravel } from "atom.io/internal"
 
-import type { AtomFamilyToken, AtomToken, TimelineToken } from "."
+import type {
+	AtomFamilyToken,
+	AtomToken,
+	TimelineToken,
+	TimelineUpdate,
+} from "."
 
 export type TimelineManageable = AtomFamilyToken<any, any> | AtomToken<any>
 export type AtomOnly<M extends TimelineManageable> = M extends AtomFamilyToken<
@@ -36,15 +32,6 @@ export const redo = (timeline: TimelineToken<any>): void => {
 export const undo = (timeline: TimelineToken<any>): void => {
 	timeTravel(IMPLICIT.STORE, `undo`, timeline)
 }
-
-export type TimelineUpdate<ManagedAtom extends TimelineManageable> =
-	| TimelineAtomUpdate<ManagedAtom>
-	| TimelineMoleculeCreation
-	| TimelineMoleculeDisposal
-	| TimelineSelectorUpdate<ManagedAtom>
-	| TimelineStateCreation<AtomOnly<ManagedAtom>>
-	| TimelineStateDisposal<AtomOnly<ManagedAtom>>
-	| TimelineTransactionUpdate
 
 export type TimelineOptions<ManagedAtom extends TimelineManageable> = {
 	/** The unique identifier of the timeline */
